Move DomUtil to pom package
