+++ liberty-starter-application/src/main/java/com/ibm/liberty/starter/pom/SetRepositoryCommand.java
@@ -10,7 +10,6 @@ import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
 import com.ibm.liberty.starter.DependencyHandler;
-import com.ibm.liberty.starter.DomUtil;
 
 public class SetRepositoryCommand implements PomModifierCommand {
