+++ liberty-starter-application/src/test/java/com/ibm/liberty/starter/pom/unit/AppNameCommandTest.java
@@ -14,8 +14,8 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 
 import com.ibm.liberty.starter.DependencyHandler;
-import com.ibm.liberty.starter.DomUtil;
 import com.ibm.liberty.starter.pom.AppNameCommand;
+import com.ibm.liberty.starter.pom.DomUtil;
 import com.ibm.liberty.starter.unit.MockDependencyHandler;
 
 public class AppNameCommandTest {
